List of 'errno' error codes

For the errors that are not used on the current platform, value is NA_integer_.

Usage

errno()

Details

A data frame with columns: name, value, description.

Examples

errno()
#>                name value
#> 1             E2BIG     7
#> 2            EACCES    13
#> 3        EADDRINUSE    98
#> 4     EADDRNOTAVAIL    99
#> 5              EADV    68
#> 6      EAFNOSUPPORT    97
#> 7            EAGAIN    11
#> 8          EALREADY   114
#> 9             EAUTH    NA
#> 10      EBACKGROUND    NA
#> 11         EBADARCH    NA
#> 12            EBADE    52
#> 13         EBADEXEC    NA
#> 14            EBADF     9
#> 15           EBADFD    77
#> 16        EBADMACHO    NA
#> 17          EBADMSG    74
#> 18            EBADR    53
#> 19          EBADRPC    NA
#> 20          EBADRQC    56
#> 21          EBADSLT    57
#> 22           EBFONT    59
#> 23            EBUSY    16
#> 24        ECANCELED   125
#> 25           ECHILD    10
#> 26           ECHRNG    44
#> 27            ECOMM    70
#> 28     ECONNABORTED   103
#> 29     ECONNREFUSED   111
#> 30       ECONNRESET   104
#> 31               ED    NA
#> 32          EDEADLK    35
#> 33        EDEADLOCK    35
#> 34     EDESTADDRREQ    89
#> 35          EDEVERR    NA
#> 36            EDIED    NA
#> 37             EDOM    33
#> 38          EDOTDOT    73
#> 39           EDQUOT   122
#> 40           EEXIST    17
#> 41           EFAULT    14
#> 42            EFBIG    27
#> 43           EFTYPE    NA
#> 44      EGRATUITOUS    NA
#> 45        EGREGIOUS    NA
#> 46        EHOSTDOWN   112
#> 47     EHOSTUNREACH   113
#> 48        EHWPOISON   133
#> 49            EIDRM    43
#> 50            EIEIO    NA
#> 51           EILSEQ    84
#> 52      EINPROGRESS   115
#> 53            EINTR     4
#> 54           EINVAL    22
#> 55              EIO     5
#> 56          EISCONN   106
#> 57           EISDIR    21
#> 58           EISNAM   120
#> 59      EKEYEXPIRED   127
#> 60     EKEYREJECTED   129
#> 61      EKEYREVOKED   128
#> 62           EL2HLT    51
#> 63         EL2NSYNC    45
#> 64           EL3HLT    46
#> 65           EL3RST    47
#> 66          ELIBACC    79
#> 67          ELIBBAD    80
#> 68         ELIBEXEC    83
#> 69          ELIBMAX    82
#> 70          ELIBSCN    81
#> 71           ELNRNG    48
#> 72            ELOOP    40
#> 73      EMEDIUMTYPE   124
#> 74           EMFILE    24
#> 75           EMLINK    31
#> 76         EMSGSIZE    90
#> 77        EMULTIHOP    72
#> 78     ENAMETOOLONG    36
#> 79          ENAVAIL   119
#> 80        ENEEDAUTH    NA
#> 81         ENETDOWN   100
#> 82        ENETRESET   102
#> 83      ENETUNREACH   101
#> 84           ENFILE    23
#> 85           ENOANO    55
#> 86          ENOATTR    NA
#> 87          ENOBUFS   105
#> 88           ENOCSI    50
#> 89          ENODATA    61
#> 90           ENODEV    19
#> 91           ENOENT     2
#> 92          ENOEXEC     8
#> 93           ENOKEY   126
#> 94           ENOLCK    37
#> 95          ENOLINK    67
#> 96        ENOMEDIUM   123
#> 97           ENOMEM    12
#> 98           ENOMSG    42
#> 99           ENONET    64
#> 100          ENOPKG    65
#> 101     ENOPROTOOPT    92
#> 102          ENOSPC    28
#> 103           ENOSR    63
#> 104          ENOSTR    60
#> 105          ENOSYS    38
#> 106         ENOTBLK    15
#> 107        ENOTCONN   107
#> 108         ENOTDIR    20
#> 109       ENOTEMPTY    39
#> 110         ENOTNAM   118
#> 111 ENOTRECOVERABLE   131
#> 112        ENOTSOCK    88
#> 113         ENOTSUP    95
#> 114          ENOTTY    25
#> 115        ENOTUNIQ    76
#> 116           ENXIO     6
#> 117      EOPNOTSUPP    95
#> 118       EOVERFLOW    75
#> 119      EOWNERDEAD   130
#> 120           EPERM     1
#> 121    EPFNOSUPPORT    96
#> 122           EPIPE    32
#> 123        EPROCLIM    NA
#> 124    EPROCUNAVAIL    NA
#> 125   EPROGMISMATCH    NA
#> 126    EPROGUNAVAIL    NA
#> 127          EPROTO    71
#> 128 EPROTONOSUPPORT    93
#> 129      EPROTOTYPE    91
#> 130         EPWROFF    NA
#> 131          ERANGE    34
#> 132         EREMCHG    78
#> 133         EREMOTE    66
#> 134       EREMOTEIO   121
#> 135        ERESTART    85
#> 136         ERFKILL   132
#> 137           EROFS    30
#> 138    ERPCMISMATCH    NA
#> 139      ESHLIBVERS    NA
#> 140       ESHUTDOWN   108
#> 141 ESOCKTNOSUPPORT    94
#> 142          ESPIPE    29
#> 143           ESRCH     3
#> 144          ESRMNT    69
#> 145          ESTALE   116
#> 146        ESTRPIPE    86
#> 147           ETIME    62
#> 148       ETIMEDOUT   110
#> 149    ETOOMANYREFS   109
#> 150         ETXTBSY    26
#> 151         EUCLEAN   117
#> 152         EUNATCH    49
#> 153          EUSERS    87
#> 154     EWOULDBLOCK    11
#> 155           EXDEV    18
#> 156          EXFULL    54
#>                                             description
#> 1                                    Arg list too long.
#> 2                                    Permission denied.
#> 3                               Address already in use.
#> 4                      Cannot assign requested address.
#> 5                                      Advertise error.
#> 6      Address family not supported by protocol family.
#> 7                     Resource temporarily unavailable.
#> 8                        Operation already in progress.
#> 9                                 Authentication error.
#> 10           Caller not in the foreground process group
#> 11                          Bad CPU type in executable.
#> 12                                    Invalid exchange.
#> 13                  Bad executable (or shared library).
#> 14                                 Bad file descriptor.
#> 15                        File descriptor in bad state.
#> 16                               Malformed Mach-o file.
#> 17                                         Bad message.
#> 18                          Invalid request descriptor.
#> 19                                   RPC struct is bad.
#> 20                                Invalid request code.
#> 21                                        Invalid slot.
#> 22                                Bad font file format.
#> 23                                       Resource busy.
#> 24                                  Operation canceled.
#> 25                                  No child processes.
#> 26                         Channel number out of range.
#> 27                         Communication error on send.
#> 28                    Software caused connection abort.
#> 29                                  Connection refused.
#> 30                            Connection reset by peer.
#> 31   The experienced user will know what is wrong.#else
#> 32                           Resource deadlock avoided.
#> 33                         File locking deadlock error.
#> 34                        Destination address required.
#> 35                                        Device error.
#> 36                                      Translator died
#> 37                    Numerical argument out of domain.
#> 38                                   RFS specific error
#> 39                                 Disc quota exceeded.
#> 40                                         File exists.
#> 41                                         Bad address.
#> 42                                      File too large.
#> 43                   Inappropriate file type or format.
#> 44                      This error code has no purpose.
#> 45                                      You did *what*?
#> 46                                        Host is down.
#> 47                                    No route to host.
#> 48                      Memory page has hardware error.
#> 49                                  Identifier removed.
#> 50  Go home and have a glass of warm, dairy-fresh milk.
#> 51                               Illegal byte sequence.
#> 52                           Operation now in progress.
#> 53                           Interrupted function call.
#> 54                                    Invalid argument.
#> 55                                  Input/output error.
#> 56                         Socket is already connected.
#> 57                                      Is a directory.
#> 58                                Is a named type file.
#> 59                                     Key has expired.
#> 60                         Key was rejected by service.
#> 61                                Key has been revoked.
#> 62                                      Level 2 halted.
#> 63                            Level 2 not synchronized.
#> 64                                      Level 3 halted.
#> 65                                       Level 3 reset.
#> 66              Can not access a needed shared library.
#> 67                Accessing a corrupted shared library.
#> 68               Cannot exec a shared library directly.
#> 69     Attempting to link in too many shared libraries.
#> 70                     .lib section in a.out corrupted.
#> 71                            Link number out of range.
#> 72                   Too many levels of symbolic links.
#> 73                                   Wrong medium type.
#> 74                                 Too many open files.
#> 75                                      Too many links.
#> 76                                    Message too long.
#> 77                                  Multihop attempted.
#> 78                                  File name too long.
#> 79                       No XENIX semaphores available.
#> 80                                  Need authenticator.
#> 81                                     Network is down.
#> 82                 Network dropped connection on reset.
#> 83                              Network is unreachable.
#> 84                       Too many open files in system.
#> 85                                            No anode.
#> 86                                 Attribute not found.
#> 87                           No buffer space available.
#> 88                          No CSI structure available.
#> 89                                No message available.
#> 90                   Operation not supported by device.
#> 91                           No such file or directory.
#> 92                                   Exec format error.
#> 93                          Required key not available.
#> 94                                  No locks available.
#> 95                               Link has been severed.
#> 96                                     No medium found.
#> 97                              Cannot allocate memory.
#> 98                          No message of desired type.
#> 99                       Machine is not on the network.
#> 100                              Package not installed.
#> 101                             Protocol not available.
#> 102                                Device out of space.
#> 103                           Out of streams resources.
#> 104                                       Not a STREAM.
#> 105                           Function not implemented.
#> 106                                 Not a block device.
#> 107                            Socket is not connected.
#> 108                                    Not a directory.
#> 109                                Directory not empty.
#> 110                        Not a XENIX named type file.
#> 111                              State not recoverable.
#> 112                     Socket operation on non-socket.
#> 113                                      Not supported.
#> 114                     Inappropriate ioctl for device.
#> 115                         Name not unique on network.
#> 116                          No such device or address.
#> 117                  Operation not supported on socket.
#> 118          Value too large to be stored in data type.
#> 119                                         Owner died.
#> 120                            Operation not permitted.
#> 121                      Protocol family not supported.
#> 122                                        Broken pipe.
#> 123                                 Too many processes.
#> 124                          Bad procedure for program.
#> 125                              Program version wrong.
#> 126                                RPC prog. not avail.
#> 127                                     Protocol error.
#> 128                             Protocol not supported.
#> 129                     Protocol wrong type for socket.
#> 130                                Device power is off.
#> 131                      Numerical result out of range.
#> 132                             Remote address changed.
#> 133                         File is already NFS-mounted
#> 134                                   Remote I/O error.
#> 135        Interrupted system call should be restarted.
#> 136              Operation not possible due to RF-kill.
#> 137                              Read-only file system.
#> 138                                  RPC version wrong.
#> 139                    Shared library version mismatch.
#> 140                  Cannot send after socket shutdown.
#> 141                          Socket type not supported.
#> 142                                       Illegal seek.
#> 143                                    No such process.
#> 144                                      Srmount error.
#> 145                              Stale NFS file handle.
#> 146                                 Streams pipe error.
#> 147                             STREAM ioctl() timeout.
#> 148                                Operation timed out.
#> 149                 Too many references: cannot splice.
#> 150                                     Text file busy.
#> 151                           Structure needs cleaning.
#> 152                       Protocol driver not attached.
#> 153                                     Too many users.
#> 154                   Resource temporarily unavailable.
#> 155                                      Improper link.
#> 156                                      Exchange full.