List of child processes (process objects) of the process. Note that this typically requires enumerating all processes on the system, so it is a costly operation.

List of child processes (process objects) of the process. Note that this typically requires enumerating all processes on the system, so it is a costly operation.

ps_children(p, recursive = FALSE)

Arguments

p
Process handle.

recursive
Whether to include the children of the children, etc.

Value

List of ps_handle objects.

See also

Examples

p <- ps_parent(ps_handle())
ps_children(p)