CPU times of the process

All times are measured in seconds:

user: Amount of time that this process has been scheduled in user mode.

system: Amount of time that this process has been scheduled in kernel mode

children_user: On Linux, amount of time that this process's waited-for children have been scheduled in user mode.

children_system: On Linux, Amount of time that this process's waited-for children have been scheduled in kernel mode.

Usage

ps_cpu_times(p = ps_handle())

Arguments

p

Process handle.

Value

Named real vector or length four: user, system, children_user, children_system. The last two are NA on non-Linux systems.

Details

Throws a zombie_process() error for zombie processes.

See also

Examples

p <- ps_handle()
p
#> <ps::ps_handle> PID=6161, NAME=R, AT=2024-09-01 08:39:15.7
ps_cpu_times(p)
#>            user          system   children_user children_system 
#>            7.26            0.61            1.10            0.89 
proc.time()
#>    user  system elapsed 
#>   8.368   1.509  12.655