ps_disk_partitions {ps}	R Documentation

List all mounted partitions

Description

The output is similar the Unix mount and df commands.

Usage

ps_disk_partitions(all = FALSE)

Arguments

all
Whether to list virtual devices as well. If FALSE, on Linux it will still list overlay and grpcfuse file systems, to provide some useful information in Docker containers.

Value

A data frame with columns device, mountpoint, fstype and options.

See Also

Other disk functions: ps_disk_usage()

Examples

ps_disk_partitions(all = TRUE)
ps_disk_partitions()

[Package ps version 1.7.5 Index]