ps_fs_mount_point {ps}	R Documentation

Find the mount point of a file or directory

Description

Find the mount point of a file or directory

Usage

ps_fs_mount_point(paths)

Arguments

paths
Paths to files, directories, devices, etc. They must exist. They are normalized using base::normalizePath().

Value

Character vector, paths to the mount points of the input paths.

Examples

ps_fs_mount_point(".")

[Package ps version 1.9.1 Index]