ps_status {ps}	R Documentation

Current process status

Description

One of the following:

"idle": Process being created by fork, macOS only.

"running": Currently runnable on macOS and Windows. Actually running on Linux.

"sleeping" Sleeping on a wait or poll.

"disk_sleep" Uninterruptible sleep, waiting for an I/O operation (Linux only).

"stopped" Stopped, either by a job control signal or because it is being traced.

"tracing_stop" Stopped for tracing (Linux only).

"zombie" Zombie. Finished, but parent has not read out the exit status yet.

"dead" Should never be seen (Linux).

"wake_kill" Received fatal signal (Linux only).

"waking" Paging (Linux only, not valid since the 2.6.xx kernel).

Usage

ps_status(p = ps_handle())

Arguments

p
Process handle.

Details

Works for zombie processes.

Value

Character scalar.

See Also

Other process handle functions: ps_children(), ps_cmdline(), ps_connections(), ps_cpu_times(), ps_create_time(), ps_cwd(), ps_descent(), ps_environ(), ps_exe(), ps_handle(), ps_interrupt(), ps_is_running(), ps_kill(), ps_memory_info(), ps_name(), ps_num_fds(), ps_num_threads(), ps_open_files(), ps_pid(), ps_ppid(), ps_resume(), ps_send_signal(), ps_shared_libs(), ps_suspend(), ps_terminal(), ps_terminate(), ps_uids(), ps_username()

Examples

p <- ps_handle()
p
ps_status(p)

[Package ps version 1.7.2 Index]