ps_system_cpu_times {ps}	R Documentation

System CPU times.

Description

Every attribute represents the seconds the CPU has spent in the given mode. The attributes availability varies depending on the platform:

user: time spent by normal processes executing in user mode; on Linux this also includes guest time.

system: time spent by processes executing in kernel mode.

idle: time spent doing nothing.

Usage

ps_system_cpu_times()

Details

Platform-specific fields:

nice (UNIX): time spent by niced (prioritized) processes executing in user mode; on Linux this also includes guest_nice time.

iowait (Linux): time spent waiting for I/O to complete. This is not accounted in idle time counter.

irq (Linux): time spent for servicing hardware interrupts.

softirq (Linux): time spent for servicing software interrupts.

steal (Linux 2.6.11+): time spent by other operating systems running in a virtualized environment.

guest (Linux 2.6.24+): time spent running a virtual CPU for guest operating systems under the control of the Linux kernel.

guest_nice (Linux 3.2.0+): time spent running a niced guest (virtual CPU for guest operating systems under the control of the Linux kernel).

Value

Named list

Examples

ps_system_cpu_times()

[Package ps version 1.7.5 Index]