ps_system_swap {ps}	R Documentation

System swap memory statistics

Description

System swap memory statistics

Usage

ps_system_swap()

Value

Named list. All numbers are in bytes:

total: total swap memory.

used: used swap memory.

free: free swap memory.

percent: the percentage usage.

sin: the number of bytes the system has swapped in from disk (cumulative). This is NA on Windows.

sout: the number of bytes the system has swapped out from disk (cumulative). This is NA on Windows.

See Also

Other memory functions: ps_system_memory()

Examples

ps_system_swap()

[Package ps version 1.7.5 Index]