List of 'errno' error codes

For the errors that are not used on the current platform, value is NA_integer_.

errno()

Details

A data frame with columns: name, value, description.

Examples

errno()

#>                name value                                         description
#> 1             E2BIG     7                                  Arg list too long.
#> 2            EACCES    13                                  Permission denied.
#> 3        EADDRINUSE    98                             Address already in use.
#> 4     EADDRNOTAVAIL    99                    Cannot assign requested address.
#> 5              EADV    68                                    Advertise error.
#> 6      EAFNOSUPPORT    97    Address family not supported by protocol family.
#> 7            EAGAIN    11                   Resource temporarily unavailable.
#> 8          EALREADY   114                      Operation already in progress.
#> 9             EAUTH    NA                               Authentication error.
#> 10      EBACKGROUND    NA          Caller not in the foreground process group
#> 11         EBADARCH    NA                         Bad CPU type in executable.
#> 12            EBADE    52                                   Invalid exchange.
#> 13         EBADEXEC    NA                 Bad executable (or shared library).
#> 14            EBADF     9                                Bad file descriptor.
#> 15           EBADFD    77                       File descriptor in bad state.
#> 16        EBADMACHO    NA                              Malformed Mach-o file.
#> 17          EBADMSG    74                                        Bad message.
#> 18            EBADR    53                         Invalid request descriptor.
#> 19          EBADRPC    NA                                  RPC struct is bad.
#> 20          EBADRQC    56                               Invalid request code.
#> 21          EBADSLT    57                                       Invalid slot.
#> 22           EBFONT    59                               Bad font file format.
#> 23            EBUSY    16                                      Resource busy.
#> 24        ECANCELED   125                                 Operation canceled.
#> 25           ECHILD    10                                 No child processes.
#> 26           ECHRNG    44                        Channel number out of range.
#> 27            ECOMM    70                        Communication error on send.
#> 28     ECONNABORTED   103                   Software caused connection abort.
#> 29     ECONNREFUSED   111                                 Connection refused.
#> 30       ECONNRESET   104                           Connection reset by peer.
#> 31               ED    NA  The experienced user will know what is wrong.#else
#> 32          EDEADLK    35                          Resource deadlock avoided.
#> 33        EDEADLOCK    35                        File locking deadlock error.
#> 34     EDESTADDRREQ    89                       Destination address required.
#> 35          EDEVERR    NA                                       Device error.
#> 36            EDIED    NA                                     Translator died
#> 37             EDOM    33                   Numerical argument out of domain.
#> 38          EDOTDOT    73                                  RFS specific error
#> 39           EDQUOT   122                                Disc quota exceeded.
#> 40           EEXIST    17                                        File exists.
#> 41           EFAULT    14                                        Bad address.
#> 42            EFBIG    27                                     File too large.
#> 43           EFTYPE    NA                  Inappropriate file type or format.
#> 44      EGRATUITOUS    NA                     This error code has no purpose.
#> 45        EGREGIOUS    NA                                     You did *what*?
#> 46        EHOSTDOWN   112                                       Host is down.
#> 47     EHOSTUNREACH   113                                   No route to host.
#> 48        EHWPOISON   133                     Memory page has hardware error.
#> 49            EIDRM    43                                 Identifier removed.
#> 50            EIEIO    NA Go home and have a glass of warm, dairy-fresh milk.
#> 51           EILSEQ    84                              Illegal byte sequence.
#> 52      EINPROGRESS   115                          Operation now in progress.
#> 53            EINTR     4                          Interrupted function call.
#> 54           EINVAL    22                                   Invalid argument.
#> 55              EIO     5                                 Input/output error.
#> 56          EISCONN   106                        Socket is already connected.
#> 57           EISDIR    21                                     Is a directory.
#> 58           EISNAM   120                               Is a named type file.
#> 59      EKEYEXPIRED   127                                    Key has expired.
#> 60     EKEYREJECTED   129                        Key was rejected by service.
#> 61      EKEYREVOKED   128                               Key has been revoked.
#> 62           EL2HLT    51                                     Level 2 halted.
#> 63         EL2NSYNC    45                           Level 2 not synchronized.
#> 64           EL3HLT    46                                     Level 3 halted.
#> 65           EL3RST    47                                      Level 3 reset.
#> 66          ELIBACC    79             Can not access a needed shared library.
#> 67          ELIBBAD    80               Accessing a corrupted shared library.
#> 68         ELIBEXEC    83              Cannot exec a shared library directly.
#> 69          ELIBMAX    82    Attempting to link in too many shared libraries.
#> 70          ELIBSCN    81                    .lib section in a.out corrupted.
#> 71           ELNRNG    48                           Link number out of range.
#> 72            ELOOP    40                  Too many levels of symbolic links.
#> 73      EMEDIUMTYPE   124                                  Wrong medium type.
#> 74           EMFILE    24                                Too many open files.
#> 75           EMLINK    31                                     Too many links.
#> 76         EMSGSIZE    90                                   Message too long.
#> 77        EMULTIHOP    72                                 Multihop attempted.
#> 78     ENAMETOOLONG    36                                 File name too long.
#> 79          ENAVAIL   119                      No XENIX semaphores available.
#> 80        ENEEDAUTH    NA                                 Need authenticator.
#> 81         ENETDOWN   100                                    Network is down.
#> 82        ENETRESET   102                Network dropped connection on reset.
#> 83      ENETUNREACH   101                             Network is unreachable.
#> 84           ENFILE    23                      Too many open files in system.
#> 85           ENOANO    55                                           No anode.
#> 86          ENOATTR    NA                                Attribute not found.
#> 87          ENOBUFS   105                          No buffer space available.
#> 88           ENOCSI    50                         No CSI structure available.
#> 89          ENODATA    61                               No message available.
#> 90           ENODEV    19                  Operation not supported by device.
#> 91           ENOENT     2                          No such file or directory.
#> 92          ENOEXEC     8                                  Exec format error.
#> 93           ENOKEY   126                         Required key not available.
#> 94           ENOLCK    37                                 No locks available.
#> 95          ENOLINK    67                              Link has been severed.
#> 96        ENOMEDIUM   123                                    No medium found.
#> 97           ENOMEM    12                             Cannot allocate memory.
#> 98           ENOMSG    42                         No message of desired type.
#> 99           ENONET    64                      Machine is not on the network.
#> 100          ENOPKG    65                              Package not installed.
#> 101     ENOPROTOOPT    92                             Protocol not available.
#> 102          ENOSPC    28                                Device out of space.
#> 103           ENOSR    63                           Out of streams resources.
#> 104          ENOSTR    60                                       Not a STREAM.
#> 105          ENOSYS    38                           Function not implemented.
#> 106         ENOTBLK    15                                 Not a block device.
#> 107        ENOTCONN   107                            Socket is not connected.
#> 108         ENOTDIR    20                                    Not a directory.
#> 109       ENOTEMPTY    39                                Directory not empty.
#> 110         ENOTNAM   118                        Not a XENIX named type file.
#> 111 ENOTRECOVERABLE   131                              State not recoverable.
#> 112        ENOTSOCK    88                     Socket operation on non-socket.
#> 113         ENOTSUP    95                                      Not supported.
#> 114          ENOTTY    25                     Inappropriate ioctl for device.
#> 115        ENOTUNIQ    76                         Name not unique on network.
#> 116           ENXIO     6                          No such device or address.
#> 117      EOPNOTSUPP    95                  Operation not supported on socket.
#> 118       EOVERFLOW    75          Value too large to be stored in data type.
#> 119      EOWNERDEAD   130                                         Owner died.
#> 120           EPERM     1                            Operation not permitted.
#> 121    EPFNOSUPPORT    96                      Protocol family not supported.
#> 122           EPIPE    32                                        Broken pipe.
#> 123        EPROCLIM    NA                                 Too many processes.
#> 124    EPROCUNAVAIL    NA                          Bad procedure for program.
#> 125   EPROGMISMATCH    NA                              Program version wrong.
#> 126    EPROGUNAVAIL    NA                                RPC prog. not avail.
#> 127          EPROTO    71                                     Protocol error.
#> 128 EPROTONOSUPPORT    93                             Protocol not supported.
#> 129      EPROTOTYPE    91                     Protocol wrong type for socket.
#> 130         EPWROFF    NA                                Device power is off.
#> 131          ERANGE    34                      Numerical result out of range.
#> 132         EREMCHG    78                             Remote address changed.
#> 133         EREMOTE    66                         File is already NFS-mounted
#> 134       EREMOTEIO   121                                   Remote I/O error.
#> 135        ERESTART    85        Interrupted system call should be restarted.
#> 136         ERFKILL   132              Operation not possible due to RF-kill.
#> 137           EROFS    30                              Read-only file system.
#> 138    ERPCMISMATCH    NA                                  RPC version wrong.
#> 139      ESHLIBVERS    NA                    Shared library version mismatch.
#> 140       ESHUTDOWN   108                  Cannot send after socket shutdown.
#> 141 ESOCKTNOSUPPORT    94                          Socket type not supported.
#> 142          ESPIPE    29                                       Illegal seek.
#> 143           ESRCH     3                                    No such process.
#> 144          ESRMNT    69                                      Srmount error.
#> 145          ESTALE   116                              Stale NFS file handle.
#> 146        ESTRPIPE    86                                 Streams pipe error.
#> 147           ETIME    62                             STREAM ioctl() timeout.
#> 148       ETIMEDOUT   110                                Operation timed out.
#> 149    ETOOMANYREFS   109                 Too many references: cannot splice.
#> 150         ETXTBSY    26                                     Text file busy.
#> 151         EUCLEAN   117                           Structure needs cleaning.
#> 152         EUNATCH    49                       Protocol driver not attached.
#> 153          EUSERS    87                                     Too many users.
#> 154     EWOULDBLOCK    11                   Resource temporarily unavailable.
#> 155           EXDEV    18                                      Improper link.
#> 156          EXFULL    54                                      Exchange full.