List of 'errno' error codes

For the errors that are not used on the current platform, value is NA_integer_.

errno()

Details

A data frame with columns: name, value, description.

Examples

errno()

#>                name value                                         description
#> 1             E2BIG     7                                  Arg list too long.
#> 2            EACCES    13                                  Permission denied.
#> 3        EADDRINUSE    48                             Address already in use.
#> 4     EADDRNOTAVAIL    49                    Cannot assign requested address.
#> 5              EADV    NA                                    Advertise error.
#> 6      EAFNOSUPPORT    47    Address family not supported by protocol family.
#> 7            EAGAIN    35                   Resource temporarily unavailable.
#> 8          EALREADY    37                      Operation already in progress.
#> 9             EAUTH    80                               Authentication error.
#> 10      EBACKGROUND    NA          Caller not in the foreground process group
#> 11         EBADARCH    86                         Bad CPU type in executable.
#> 12            EBADE    NA                                   Invalid exchange.
#> 13         EBADEXEC    85                 Bad executable (or shared library).
#> 14            EBADF     9                                Bad file descriptor.
#> 15           EBADFD    NA                       File descriptor in bad state.
#> 16        EBADMACHO    88                              Malformed Mach-o file.
#> 17          EBADMSG    94                                        Bad message.
#> 18            EBADR    NA                         Invalid request descriptor.
#> 19          EBADRPC    72                                  RPC struct is bad.
#> 20          EBADRQC    NA                               Invalid request code.
#> 21          EBADSLT    NA                                       Invalid slot.
#> 22           EBFONT    NA                               Bad font file format.
#> 23            EBUSY    16                                      Resource busy.
#> 24        ECANCELED    89                                 Operation canceled.
#> 25           ECHILD    10                                 No child processes.
#> 26           ECHRNG    NA                        Channel number out of range.
#> 27            ECOMM    NA                        Communication error on send.
#> 28     ECONNABORTED    53                   Software caused connection abort.
#> 29     ECONNREFUSED    61                                 Connection refused.
#> 30       ECONNRESET    54                           Connection reset by peer.
#> 31               ED    NA  The experienced user will know what is wrong.#else
#> 32          EDEADLK    11                          Resource deadlock avoided.
#> 33        EDEADLOCK    NA                        File locking deadlock error.
#> 34     EDESTADDRREQ    39                       Destination address required.
#> 35          EDEVERR    83                                       Device error.
#> 36            EDIED    NA                                     Translator died
#> 37             EDOM    33                   Numerical argument out of domain.
#> 38          EDOTDOT    NA                                  RFS specific error
#> 39           EDQUOT    69                                Disc quota exceeded.
#> 40           EEXIST    17                                        File exists.
#> 41           EFAULT    14                                        Bad address.
#> 42            EFBIG    27                                     File too large.
#> 43           EFTYPE    79                  Inappropriate file type or format.
#> 44      EGRATUITOUS    NA                     This error code has no purpose.
#> 45        EGREGIOUS    NA                                     You did *what*?
#> 46        EHOSTDOWN    64                                       Host is down.
#> 47     EHOSTUNREACH    65                                   No route to host.
#> 48        EHWPOISON    NA                     Memory page has hardware error.
#> 49            EIDRM    90                                 Identifier removed.
#> 50            EIEIO    NA Go home and have a glass of warm, dairy-fresh milk.
#> 51           EILSEQ    92                              Illegal byte sequence.
#> 52      EINPROGRESS    36                          Operation now in progress.
#> 53            EINTR     4                          Interrupted function call.
#> 54           EINVAL    22                                   Invalid argument.
#> 55              EIO     5                                 Input/output error.
#> 56          EISCONN    56                        Socket is already connected.
#> 57           EISDIR    21                                     Is a directory.
#> 58           EISNAM    NA                               Is a named type file.
#> 59      EKEYEXPIRED    NA                                    Key has expired.
#> 60     EKEYREJECTED    NA                        Key was rejected by service.
#> 61      EKEYREVOKED    NA                               Key has been revoked.
#> 62           EL2HLT    NA                                     Level 2 halted.
#> 63         EL2NSYNC    NA                           Level 2 not synchronized.
#> 64           EL3HLT    NA                                     Level 3 halted.
#> 65           EL3RST    NA                                      Level 3 reset.
#> 66          ELIBACC    NA             Can not access a needed shared library.
#> 67          ELIBBAD    NA               Accessing a corrupted shared library.
#> 68         ELIBEXEC    NA              Cannot exec a shared library directly.
#> 69          ELIBMAX    NA    Attempting to link in too many shared libraries.
#> 70          ELIBSCN    NA                    .lib section in a.out corrupted.
#> 71           ELNRNG    NA                           Link number out of range.
#> 72            ELOOP    62                  Too many levels of symbolic links.
#> 73      EMEDIUMTYPE    NA                                  Wrong medium type.
#> 74           EMFILE    24                                Too many open files.
#> 75           EMLINK    31                                     Too many links.
#> 76         EMSGSIZE    40                                   Message too long.
#> 77        EMULTIHOP    95                                 Multihop attempted.
#> 78     ENAMETOOLONG    63                                 File name too long.
#> 79          ENAVAIL    NA                      No XENIX semaphores available.
#> 80        ENEEDAUTH    81                                 Need authenticator.
#> 81         ENETDOWN    50                                    Network is down.
#> 82        ENETRESET    52                Network dropped connection on reset.
#> 83      ENETUNREACH    51                             Network is unreachable.
#> 84           ENFILE    23                      Too many open files in system.
#> 85           ENOANO    NA                                           No anode.
#> 86          ENOATTR    93                                Attribute not found.
#> 87          ENOBUFS    55                          No buffer space available.
#> 88           ENOCSI    NA                         No CSI structure available.
#> 89          ENODATA    96                               No message available.
#> 90           ENODEV    19                  Operation not supported by device.
#> 91           ENOENT     2                          No such file or directory.
#> 92          ENOEXEC     8                                  Exec format error.
#> 93           ENOKEY    NA                         Required key not available.
#> 94           ENOLCK    77                                 No locks available.
#> 95          ENOLINK    97                              Link has been severed.
#> 96        ENOMEDIUM    NA                                    No medium found.
#> 97           ENOMEM    12                             Cannot allocate memory.
#> 98           ENOMSG    91                         No message of desired type.
#> 99           ENONET    NA                      Machine is not on the network.
#> 100          ENOPKG    NA                              Package not installed.
#> 101     ENOPROTOOPT    42                             Protocol not available.
#> 102          ENOSPC    28                                Device out of space.
#> 103           ENOSR    98                           Out of streams resources.
#> 104          ENOSTR    99                                       Not a STREAM.
#> 105          ENOSYS    78                           Function not implemented.
#> 106         ENOTBLK    15                                 Not a block device.
#> 107        ENOTCONN    57                            Socket is not connected.
#> 108         ENOTDIR    20                                    Not a directory.
#> 109       ENOTEMPTY    66                                Directory not empty.
#> 110         ENOTNAM    NA                        Not a XENIX named type file.
#> 111 ENOTRECOVERABLE   104                              State not recoverable.
#> 112        ENOTSOCK    38                     Socket operation on non-socket.
#> 113         ENOTSUP    45                                      Not supported.
#> 114          ENOTTY    25                     Inappropriate ioctl for device.
#> 115        ENOTUNIQ    NA                         Name not unique on network.
#> 116           ENXIO     6                          No such device or address.
#> 117      EOPNOTSUPP   102                  Operation not supported on socket.
#> 118       EOVERFLOW    84          Value too large to be stored in data type.
#> 119      EOWNERDEAD   105                                         Owner died.
#> 120           EPERM     1                            Operation not permitted.
#> 121    EPFNOSUPPORT    46                      Protocol family not supported.
#> 122           EPIPE    32                                        Broken pipe.
#> 123        EPROCLIM    67                                 Too many processes.
#> 124    EPROCUNAVAIL    76                          Bad procedure for program.
#> 125   EPROGMISMATCH    75                              Program version wrong.
#> 126    EPROGUNAVAIL    74                                RPC prog. not avail.
#> 127          EPROTO   100                                     Protocol error.
#> 128 EPROTONOSUPPORT    43                             Protocol not supported.
#> 129      EPROTOTYPE    41                     Protocol wrong type for socket.
#> 130         EPWROFF    82                                Device power is off.
#> 131          ERANGE    34                      Numerical result out of range.
#> 132         EREMCHG    NA                             Remote address changed.
#> 133         EREMOTE    71                         File is already NFS-mounted
#> 134       EREMOTEIO    NA                                   Remote I/O error.
#> 135        ERESTART    NA        Interrupted system call should be restarted.
#> 136         ERFKILL    NA              Operation not possible due to RF-kill.
#> 137           EROFS    30                              Read-only file system.
#> 138    ERPCMISMATCH    73                                  RPC version wrong.
#> 139      ESHLIBVERS    87                    Shared library version mismatch.
#> 140       ESHUTDOWN    58                  Cannot send after socket shutdown.
#> 141 ESOCKTNOSUPPORT    44                          Socket type not supported.
#> 142          ESPIPE    29                                       Illegal seek.
#> 143           ESRCH     3                                    No such process.
#> 144          ESRMNT    NA                                      Srmount error.
#> 145          ESTALE    70                              Stale NFS file handle.
#> 146        ESTRPIPE    NA                                 Streams pipe error.
#> 147           ETIME   101                             STREAM ioctl() timeout.
#> 148       ETIMEDOUT    60                                Operation timed out.
#> 149    ETOOMANYREFS    59                 Too many references: cannot splice.
#> 150         ETXTBSY    26                                     Text file busy.
#> 151         EUCLEAN    NA                           Structure needs cleaning.
#> 152         EUNATCH    NA                       Protocol driver not attached.
#> 153          EUSERS    68                                     Too many users.
#> 154     EWOULDBLOCK    35                   Resource temporarily unavailable.
#> 155           EXDEV    18                                      Improper link.
#> 156          EXFULL    NA                                      Exchange full.