ps {ps}	R Documentation

Process table

Description

Process table

Usage

ps(user = NULL, after = NULL)

Arguments

user
Username, to filter the results to matching processes.

after
Start time (POSIXt), to filter the results to processes that started after this.

Value

Data frame (tibble), see columns below.

Columns:

pid: Process ID.

ppid: Process ID of parent process.

name: Process name.

username: Name of the user (real uid on POSIX).

status: I.e. running, sleeping, etc.

user: User CPU time.

system: System CPU time.

rss: Resident set size, the amount of memory the process currently uses. Does not include memory that is swapped out. It does include shared libraries.

vms: Virtual memory size. All memory the process has access to.

created: Time stamp when the process was created.

ps_handle: ps_handle objects, in a list column.

[Package ps version 1.3.0 Index]