List all mounted partitions

The output is similar the Unix mount and df commands.

ps_disk_partitions(all = FALSE)

Arguments

all
Whether to list virtual devices as well. If FALSE, on Linux it will still list overlay and grpcfuse file systems, to provide some useful information in Docker containers.

Value

A data frame (tibble) with columns device, mountpoint, fstype and options.

See also

Examples

ps_disk_partitions(all = TRUE)

#> # A tibble: 6 x 4
#>   device              mountpoint            fstype options                      
#>   <chr>               <chr>                 <chr>  <chr>                        
#> 1 /dev/disk1s5        /                     apfs   ro,local,rootfs,dovolfs,jour…
#> 2 devfs               /dev                  devfs  rw,local,dontbrowse,multilab…
#> 3 /dev/disk1s1        /System/Volumes/Data  apfs   rw,local,dovolfs,dontbrowse,…
#> 4 /dev/disk1s4        /private/var/vm       apfs   rw,local,dovolfs,dontbrowse,…
#> 5 map auto_home       /System/Volumes/Data… autofs rw,dontbrowse,automounted,mu…
#> 6 .host:/VMware Shar… /Volumes/VMware Shar… vmhgfs rw                           

ps_disk_partitions()

#> # A tibble: 3 x 4
#>   device      mountpoint         fstype options                                 
#> * <chr>       <chr>              <chr>  <chr>                                   
#> 1 /dev/disk1… /                  apfs   ro,local,rootfs,dovolfs,journaled,multi…
#> 2 /dev/disk1… /System/Volumes/D… apfs   rw,local,dovolfs,dontbrowse,journaled,m…
#> 3 /dev/disk1… /private/var/vm    apfs   rw,local,dovolfs,dontbrowse,journaled,m…