ps_disk_usage {ps}	R Documentation

Disk usage statistics, per partition

Description

The output is similar to the Unix df command.

Usage

ps_disk_usage(paths = ps_disk_partitions()$mountpoint)

Arguments

paths
The mounted file systems to list. By default all file systems returned by ps_disk_partitions() is listed.

Details

Note that on Unix a small percentage of the disk space (5% typically) is reserved for the superuser. ps_disk_usage() returns the space available to the calling user.

Value

A data frame with columns mountpoint, total, used, available and capacity.

See Also

Other disk functions: ps_disk_partitions()

Examples

ps_disk_usage()

[Package ps version 1.7.1 Index]