Disk usage statistics, per partition

The output is similar to the Unix df command.

ps_disk_usage(paths = ps_disk_partitions()$mountpoint)

Arguments

paths
The mounted file systems to list. By default all file systems returned by ps_disk_partitions() is listed.

Value

A data frame with columns mountpoint, total, used, available and capacity.

Details

Note that on Unix a small percentage of the disk space (5% typically) is reserved for the superuser. ps_disk_usage() returns the space available to the calling user.

See also

Examples

ps_disk_usage()

#> # A tibble: 3 x 5
#>   mountpoint                  total         used    available capacity
#>   <chr>                       <dbl>        <dbl>        <dbl>    <dbl>
#> 1 /                    407812136960 278278385664 396496146432    0.412
#> 2 /System/Volumes/Data 407812136960 278278385664 141615972352    0.663
#> 3 /private/var/vm      407812136960 278278385664 407811067904    0.406