Disk usage statistics, per partition

The output is similar to the Unix df command.

Usage

ps_disk_usage(paths = ps_disk_partitions()$mountpoint)

Arguments

paths

The mounted file systems to list. By default all file systems returned by ps_disk_partitions() is listed.

Value

A data frame with columns mountpoint, total, used, available and capacity.

Details

Note that on Unix a small percentage of the disk space (5% typically) is reserved for the superuser. ps_disk_usage() returns the space available to the calling user.

See also

Examples

ps_disk_usage()
#> # A data frame: 6 × 5
#>   mountpoint              total        used   available capacity
#>   <chr>                   <dbl>       <dbl>       <dbl>    <dbl>
#> 1 /                 77851254784 55743123456 22124908544   0.716 
#> 2 /snap/core20/2318    67108864    67108864           0   1     
#> 3 /snap/lxd/29351      91357184    91357184           0   1     
#> 4 /snap/snapd/21759    40763392    40763392           0   1     
#> 5 /boot/efi           109395456     6338048   103057408   0.0579
#> 6 /mnt              78673350656  8338202624 74378350592   0.101