ps_environ {ps}	R Documentation

Environment variables of a process

Description

ps_environ() returns the environment variables of the process, in a named vector, similarly to the return value of Sys.getenv() (without arguments).

Usage

ps_environ(p)

ps_environ_raw(p)

Arguments

p
Process handle.

Details

Note: this usually does not reflect changes made after the process started.

ps_environ_raw() is similar to p$environ() but returns the unparsed "var=value" strings. This is faster, and sometimes good enough.

These functions throw a zombie_process error for zombie processes.

Value

ps_environ() returns a named character vector (that has a Dlist class, so it is printed nicely), ps_environ_raw() returns a character vector.

Examples

p <- ps_handle()
p
env <- ps_environ(p)
env[["R_HOME"]]

See Also

Other process handle functions: ps_children, ps_cmdline, ps_connections, ps_cpu_times, ps_create_time, ps_cwd, ps_exe, ps_handle, ps_interrupt, ps_is_running, ps_kill, ps_memory_info, ps_name, ps_num_fds, ps_num_threads, ps_open_files, ps_pid, ps_ppid, ps_resume, ps_send_signal, ps_status, ps_suspend, ps_terminal, ps_terminate, ps_uids, ps_username

Other process handle functions: ps_children, ps_cmdline, ps_connections, ps_cpu_times, ps_create_time, ps_cwd, ps_exe, ps_handle, ps_interrupt, ps_is_running, ps_kill, ps_memory_info, ps_name, ps_num_fds, ps_num_threads, ps_open_files, ps_pid, ps_ppid, ps_resume, ps_send_signal, ps_status, ps_suspend, ps_terminal, ps_terminate, ps_uids, ps_username

[Package ps version 1.3.0 Index]