Environment variables of a process

ps_environ() returns the environment variables of the process, in a named vector, similarly to the return value of Sys.getenv() (without arguments).

Usage

ps_environ(p = ps_handle())

ps_environ_raw(p = ps_handle())

Arguments

p

Process handle.

Value

ps_environ() returns a named character vector (that has a Dlist class, so it is printed nicely), ps_environ_raw() returns a character vector.

Details

Note: this usually does not reflect changes made after the process started.

ps_environ_raw() is similar to p$environ() but returns the unparsed "var=value" strings. This is faster, and sometimes good enough.

These functions throw a zombie_process error for zombie processes.

See also

Examples

p <- ps_handle()
p
#> <ps::ps_handle> PID=5793, NAME=R, AT=2023-04-19 08:49:06
env <- ps_environ(p)
env[["R_HOME"]]
#> [1] "/opt/R/4.2.3/lib/R"