Memory usage information

Memory usage information

Usage

ps_memory_info(p = ps_handle())

ps_memory_full_info(p = ps_handle())

Arguments

p

Process handle.

Value

Named real vector.

Details

ps_memory_info() returns information about memory usage.

It returns a named vector. Portable fields:

rss: "Resident Set Size", this is the non-swapped physical memory a process has used (bytes). On UNIX it matches "top"‘s 'RES' column (see doc). On Windows this is an alias for wset field and it matches "Memory" column of taskmgr.exe.

vmem: "Virtual Memory Size", this is the total amount of virtual memory used by the process (bytes). On UNIX it matches "top"‘s 'VIRT' column (see doc). On Windows this is an alias for the pagefile field and it matches the "Working set (memory)" column of taskmgr.exe.

Non-portable fields:

shared: (Linux) memory that could be potentially shared with other processes (bytes). This matches "top"‘s 'SHR' column (see doc).

text: (Linux): aka 'TRS' (text resident set) the amount of memory devoted to executable code (bytes). This matches "top"‘s 'CODE' column (see doc).

data: (Linux): aka 'DRS' (data resident set) the amount of physical memory devoted to other than executable code (bytes). It matches "top"‘s 'DATA' column (see doc).

lib: (Linux): the memory used by shared libraries (bytes).

dirty: (Linux): the amount of memory in dirty pages (bytes).

pfaults: (macOS): number of page faults.

pageins: (macOS): number of actual pageins.

For the explanation of Windows fields see the PROCESS_MEMORY_COUNTERS_EX structure.

ps_memory_full_info() returns all fields as ps_memory_info(), plus additional information, but typically takes slightly longer to run, and might not have access to some processes that ps_memory_info() can query:

uss: Unique Set Size, this is the memory which is unique to a process and which would be freed if the process was terminated right now.

pss (Linux only): Proportional Set Size, is the amount of memory shared with other processes, accounted in a way that the amount is divided evenly between the processes that share it. I.e. if a process has 10 MBs all to itself and 10 MBs shared with another process its PSS will be 15 MBs.

swap (Linux only): amount of memory that has been swapped out to disk.

They both throw a zombie_process() error for zombie processes.

See also

Examples

p <- ps_handle()
p
#> <ps::ps_handle> PID=5992, NAME=R, AT=2024-08-29 11:23:38.29
ps_memory_info(p)
#>       rss       vms    shared      text       lib      data     dirty 
#> 260358144 951169024  33685504      4096         0 751423488         0 
ps_memory_full_info(p)
#>       rss       vms    shared      text       lib      data     dirty 
#> 260358144 951169024  33685504      4096         0 751423488         0 
#>       uss       pss      swap 
#> 241823744 245556224         0