Memory usage information

A list with information about memory usage. Portable fields:

rss: "Resident Set Size", this is the non-swapped physical memory a process has used. On UNIX it matches "top"‘s 'RES' column (see doc). On Windows this is an alias for wset field and it matches "Memory" column of taskmgr.exe.

vmem: "Virtual Memory Size", this is the total amount of virtual memory used by the process. On UNIX it matches "top"‘s 'VIRT' column (see doc). On Windows this is an alias for the pagefile field and it matches the "Working set (memory)" column of taskmgr.exe.

ps_memory_info(p = ps_handle())

Arguments

p
Process handle.

Value

Named real vector.

Details

Non-portable fields:

shared: (Linux) memory that could be potentially shared with other processes. This matches "top"‘s 'SHR' column (see doc).

text: (Linux): aka 'TRS' (text resident set) the amount of memory devoted to executable code. This matches "top"‘s 'CODE' column (see doc).

data: (Linux): aka 'DRS' (data resident set) the amount of physical memory devoted to other than executable code. It matches "top"‘s 'DATA' column (see doc).

lib: (Linux): the memory used by shared libraries.

dirty: (Linux): the number of dirty pages.

pfaults: (macOS): number of page faults.

pageins: (macOS): number of actual pageins.

For on explanation of Windows fields see the PROCESS_MEMORY_COUNTERS_EX structure.

Throws a zombie_process() error for zombie processes.

See also

Examples

p <- ps_handle()
p

#> <ps::ps_handle> PID=2252, NAME=R, AT=2020-10-06 17:14:06

ps_memory_info(p)

#>        rss        vms    pfaults    pageins 
#>  130207744 4545110016      67772          1