Open files of a process

Note that in some IDEs, e.g. RStudio or R.app on macOS, the IDE itself opens files from other threads, in addition to the files opened from the main R thread.

ps_open_files(p)

Arguments

p
Process handle.

Value

Data frame, or tibble if the tibble package is available, with columns: fd and path. fd is numeric file descriptor on POSIX systems, NA on Windows. path is an absolute path to the file.

Details

For a zombie process it throws a zombie_process error.

See also

Examples

p <- ps_handle()
ps_open_files(p)

#> # A tibble: 1 x 2
#>      fd path                                                         
#>   <int> <chr>                                                        
#> 1     3 /Users/runner/work/_temp/bbd7fd1e-e865-4631-b682-af00f9fee365

f <- file(tmp <- tempfile(), "w")
ps_open_files(p)

#> # A tibble: 2 x 2
#>      fd path                                                                    
#>   <int> <chr>                                                                   
#> 1     3 /Users/runner/work/_temp/bbd7fd1e-e865-4631-b682-af00f9fee365           
#> 2    19 /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpP9Pkij/fil…

close(f)
unlink(tmp)
ps_open_files(p)

#> # A tibble: 1 x 2
#>      fd path                                                         
#>   <int> <chr>                                                        
#> 1     3 /Users/runner/work/_temp/bbd7fd1e-e865-4631-b682-af00f9fee365