Open files of a process

Note that in some IDEs, e.g. RStudio or R.app on macOS, the IDE itself opens files from other threads, in addition to the files opened from the main R thread.

Usage

ps_open_files(p = ps_handle())

Arguments

p

Process handle.

Value

Data frame with columns: fd and path. fd is numeric file descriptor on POSIX systems, NA on Windows. path is an absolute path to the file.

Details

For a zombie process it throws a zombie_process error.

See also

Examples

p <- ps_handle()
ps_open_files(p)
#> # A data frame: 12 × 2
#>       fd path                                                        
#>    <int> <chr>                                                       
#>  1     0 pipe:[44451]                                                
#>  2     1 pipe:[44452]                                                
#>  3     2 pipe:[44453]                                                
#>  4     3 /home/runner/work/_temp/515b0e45-ca34-4be1-9477-013225b0160f
#>  5     4 anon_inode:[eventpoll]                                      
#>  6     5 pipe:[40827]                                                
#>  7     6 pipe:[40827]                                                
#>  8     7 pipe:[40828]                                                
#>  9     8 pipe:[40828]                                                
#> 10     9 anon_inode:[eventfd]                                        
#> 11   110 pipe:[25890]                                                
#> 12   113 pipe:[25891]                                                
f <- file(tmp <- tempfile(), "w")
ps_open_files(p)
#> # A data frame: 13 × 2
#>       fd path                                                        
#>    <int> <chr>                                                       
#>  1     0 pipe:[44451]                                                
#>  2     1 pipe:[44452]                                                
#>  3     2 pipe:[44453]                                                
#>  4     3 /home/runner/work/_temp/515b0e45-ca34-4be1-9477-013225b0160f
#>  5     4 anon_inode:[eventpoll]                                      
#>  6     5 pipe:[40827]                                                
#>  7     6 pipe:[40827]                                                
#>  8     7 pipe:[40828]                                                
#>  9     8 pipe:[40828]                                                
#> 10     9 anon_inode:[eventfd]                                        
#> 11    16 /tmp/RtmpXnNfLP/file17d869b42751                            
#> 12   110 pipe:[25890]                                                
#> 13   113 pipe:[25891]                                                
close(f)
unlink(tmp)
ps_open_files(p)
#> # A data frame: 12 × 2
#>       fd path                                                        
#>    <int> <chr>                                                       
#>  1     0 pipe:[44451]                                                
#>  2     1 pipe:[44452]                                                
#>  3     2 pipe:[44453]                                                
#>  4     3 /home/runner/work/_temp/515b0e45-ca34-4be1-9477-013225b0160f
#>  5     4 anon_inode:[eventpoll]                                      
#>  6     5 pipe:[40827]                                                
#>  7     6 pipe:[40827]                                                
#>  8     7 pipe:[40828]                                                
#>  9     8 pipe:[40828]                                                
#> 10     9 anon_inode:[eventfd]                                        
#> 11   110 pipe:[25890]                                                
#> 12   113 pipe:[25891]