Open files of a process

Note that in some IDEs, e.g. RStudio or R.app on macOS, the IDE itself opens files from other threads, in addition to the files opened from the main R thread.

Usage

ps_open_files(p = ps_handle())

Arguments

p

Process handle.

Value

Data frame, or tibble if the tibble package is available, with columns: fd and path. fd is numeric file descriptor on POSIX systems, NA on Windows. path is an absolute path to the file.

Details

For a zombie process it throws a zombie_process error.

See also

Examples

p <- ps_handle()
ps_open_files(p)
#> # A tibble: 12 × 2
#>       fd path                                                        
#>    <int> <chr>                                                       
#>  1     0 pipe:[61375]                                                
#>  2     1 pipe:[61376]                                                
#>  3     2 pipe:[61377]                                                
#>  4     3 /home/runner/work/_temp/8f35db0a-a3bd-46e3-b91a-f35f83f3f026
#>  5     4 anon_inode:[eventpoll]                                      
#>  6     5 pipe:[61389]                                                
#>  7     6 pipe:[61389]                                                
#>  8     7 pipe:[61390]                                                
#>  9     9 pipe:[61390]                                                
#> 10    10 anon_inode:[eventfd]                                        
#> 11   112 pipe:[32218]                                                
#> 12   115 pipe:[32219]                                                
f <- file(tmp <- tempfile(), "w")
ps_open_files(p)
#> # A tibble: 13 × 2
#>       fd path                                                        
#>    <int> <chr>                                                       
#>  1     0 pipe:[61375]                                                
#>  2     1 pipe:[61376]                                                
#>  3     2 pipe:[61377]                                                
#>  4     3 /home/runner/work/_temp/8f35db0a-a3bd-46e3-b91a-f35f83f3f026
#>  5     4 anon_inode:[eventpoll]                                      
#>  6     5 pipe:[61389]                                                
#>  7     6 pipe:[61389]                                                
#>  8     7 pipe:[61390]                                                
#>  9     9 pipe:[61390]                                                
#> 10    10 anon_inode:[eventfd]                                        
#> 11    15 /tmp/Rtmpp31TK7/file2a9d18979a8                             
#> 12   112 pipe:[32218]                                                
#> 13   115 pipe:[32219]                                                
close(f)
unlink(tmp)
ps_open_files(p)
#> # A tibble: 12 × 2
#>       fd path                                                        
#>    <int> <chr>                                                       
#>  1     0 pipe:[61375]                                                
#>  2     1 pipe:[61376]                                                
#>  3     2 pipe:[61377]                                                
#>  4     3 /home/runner/work/_temp/8f35db0a-a3bd-46e3-b91a-f35f83f3f026
#>  5     4 anon_inode:[eventpoll]                                      
#>  6     5 pipe:[61389]                                                
#>  7     6 pipe:[61389]                                                
#>  8     7 pipe:[61390]                                                
#>  9     9 pipe:[61390]                                                
#> 10    10 anon_inode:[eventfd]                                        
#> 11   112 pipe:[32218]                                                
#> 12   115 pipe:[32219]