ps_ppid {ps}	R Documentation

Parent pid or parent process of a process

Description

ps_ppid() returns the parent pid, ps_parent() returns a ps_handle of the parent.

Usage

ps_ppid(p)

ps_parent(p)

Arguments

p
Process handle.

Details

On POSIX systems, if the parent process terminates, another process (typically the pid 1 process) is marked as parent. ps_ppid() and ps_parent() will return this process then.

Both ps_ppid() and ps_parent() work for zombie processes.

Value

ps_ppid() returns and integer scalar, the pid of the parent of p. ps_parent() returns a ps_handle.

Examples

p <- ps_handle()
p
ps_ppid(p)
ps_parent(p)

See Also

Other process handle functions: ps_children, ps_cmdline, ps_connections, ps_cpu_times, ps_create_time, ps_cwd, ps_environ, ps_exe, ps_handle, ps_interrupt, ps_is_running, ps_kill, ps_memory_info, ps_name, ps_num_fds, ps_num_threads, ps_open_files, ps_pid, ps_resume, ps_send_signal, ps_status, ps_suspend, ps_terminal, ps_terminate, ps_uids, ps_username

Other process handle functions: ps_children, ps_cmdline, ps_connections, ps_cpu_times, ps_create_time, ps_cwd, ps_environ, ps_exe, ps_handle, ps_interrupt, ps_is_running, ps_kill, ps_memory_info, ps_name, ps_num_fds, ps_num_threads, ps_open_files, ps_pid, ps_resume, ps_send_signal, ps_status, ps_suspend, ps_terminal, ps_terminate, ps_uids, ps_username

[Package ps version 1.3.0 Index]