Encode a ps_handle as a short string

A convenient format for passing between processes, naming semaphores, or using as a directory/file name. Will always be 14 alphanumeric characters, with the first and last characters guarantied to be letters. Encodes the pid and creation time for a process.

Usage

ps_string(p = ps_handle())

Arguments

p

Process handle.

Value

A process string (scalar character), that can be passed to ps_handle() in place of a pid.

Examples

(p <- ps_handle())
#> <ps::ps_handle> PID=8082, NAME=R, AT=2025-10-03 06:06:05.25
(str <- ps_string(p))
#> [1] "acZwb5eI5Nbdcq"
ps_handle(pid = str)
#> <ps::ps_handle> PID=8082, NAME=R, AT=2025-10-03 06:06:05.25