ps_users {ps}	R Documentation

List users connected to the system

Description

List users connected to the system

Usage

ps_users()

Value

A data frame (tibble) with columns username, tty, hostname, start_time, pid. tty and pid are NA on Windows. pid is the process id of the login process. For local users the hostname column is the empty string.

[Package ps version 1.3.0 Index]